Add functions to load images, flatten and unflatten
Adds the function load_images which searches the directory and all
subdirectories for images and formats the images into a flatten
vector and appends them to an array. Adds flatten_image which makes
and image matrix a 1xN vector. Adds Unflatten which turns a 1xN
vector back into a image.