Add project_image and reverse_projection functions
Add a function to represent an image in the subspace and another
function to return it to the image space. Fixed some problems with
the flatten and unflatten functions as well as related issues in the
load_image and load_images functions. Replaces append with hstack
and flatten with reshape.